requirements/base.txt
# Django core
Django>=5.2,<5.3

# Django REST Framework
djangorestframework>=3.15,<4.0

# JWT authentication
djangorestframework-simplejwt>=5.3,<6.0

# Environment variables
django-environ>=0.11,<1.0

# PostgreSQL adapter
psycopg2-binary>=2.9,<3.0

# Celery + Redis
celery>=5.4,<6.0
redis>=5.0,<6.0

# Cloud storage (S3-compatible)
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# Image processing
Pillow>=10.0,<11.0

# Filtering
django-filter>=24.0,<25.0

# OpenAPI / Swagger docs
drf-spectacular>=0.27,<1.0

# CORS headers
django-cors-headers>=4.0,<5.0
